+++ frontend_tests/node_tests/user_events.js
@@ -57,6 +57,7 @@ const me = {
     user_id: 30,
     full_name: "Me Myself",
     is_admin: true,
+    role: settings_config.user_role_values.member.code,
 };
 
 function initialize() {
@@ -117,6 +118,26 @@ run_test("updates", () => {
     assert.equal(person.is_admin, true);
     assert.equal(person.is_owner, true);
     assert.equal(person.role, settings_config.user_role_values.owner.code);
+
+    user_events.update_person({user_id: me.user_id, is_billing_admin: true});
+    person = people.get_by_email(me.email);
+    assert(person.is_billing_admin);
+    assert.equal(person.role, settings_config.user_role_values.member.code);
+    assert(page_params.is_billing_admin);
+
+    user_events.update_person({user_id: me.user_id, is_billing_admin: false});
+    person = people.get_by_email(me.email);
+    assert.equal(person.user_id, me.user_id);
+    assert(!person.is_billing_admin);
+    assert.equal(person.role, settings_config.user_role_values.member.code);
+    assert(!page_params.is_billing_admin);
+
+    user_events.update_person({user_id: isaac.user_id, is_billing_admin: false});
+    person = people.get_by_email(isaac.email);
+    assert.equal(person.user_id, isaac.user_id);
+    assert(!person.is_billing_admin);
+    assert.equal(person.role, settings_config.user_role_values.owner.code);
+    assert(!page_params.is_billing_admin);
 
     let user_id;
     let full_name;