+++ zerver/lib/actions.py
@@ -4562,6 +4562,10 @@ def do_change_user_role(
 def do_make_user_billing_admin(user_profile: UserProfile) -> None:
     user_profile.is_billing_admin = True
     user_profile.save(update_fields=["is_billing_admin"])
+    event = dict(
+        type="realm_user", op="update", person=dict(user_id=user_profile.id, is_billing_admin=True)
+    )
+    send_event(user_profile.realm, event, active_user_ids(user_profile.realm_id))
 
 
 def do_change_can_forge_sender(user_profile: UserProfile, value: bool) -> None: